refactor: reorg StreamManager.lua to Player/
