+++ TTS_xwing/src/Global.lua
@@ -110,7 +110,7 @@ HotkeyModule = require("TTS_xwing.src.Player.HotkeyManager")
 PlayerEventModule = require("TTS_xwing.src.PlayerActionModule")
 DialManagerModule = require("TTS_xwing.src.Dial.DialManager")
 ShipProximityHider = require("TTS_xwing.src.ShipProximityHider")
-StreamManagerModule = require("TTS_xwing.src.StreamManager")
+StreamManagerModule = require("TTS_xwing.src.Player.StreamManager")
 --StreamManagerModule.SetUI(self)
 --StreamManagerModule.AddPresentation("Purple", "left")
 --StreamManagerModule.AddPresentation("Blue", "right")
